requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
h11==0.12.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
six==1.16.0
SQLAlchemy==1.4.23
waitress==2.0.0
Werkzeug==2.0.1
